% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot incidents from a single year on state map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Serial number of state - integer}

\item{year}{year to plot - integer}
}
\value{
plot with state map outlined and incident location shown as points on the map
}
\description{
This function genreates a plot showing the location of each incident in the defined year and state
}
\note{
ERROR if following conditions not met: Files must be in working directory and filenames must be of format "accident_<year>.csv.bz2"
}
\examples{

\dontrun{
fars_map_state(1, 2013)
}

}
