% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read}
\alias{fars_read}
\title{Read in fars data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{character string of full filename with extension}
}
\value{
dataframe of filename data
}
\description{
This function reads and returns a dataframe of the specified file
}
\note{
ERROR if following conditions not met: Input must be character string with full extension and file location must be correct
}
\examples{
\dontrun{
fars_read("accident_2013.csv.bz2")
}

}
