% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read in multiple fars files retuning list of dataframes}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{numeric vector defining years to read in from raw file}
}
\value{
dataframes based on each yearly raw fars file
}
\description{
This function returns individual data frames from raw fars files
}
\note{
ERROR if following conditions not met: Files must be in working directory and filenames must be of format "accident_<year>.csv.bz2"
}
\examples{
\dontrun{
fars_read_years(2013:2015)
}

}
