% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{make_filename}
\alias{make_filename}
\title{Generate file name string from year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{integer of year}
}
\value{
string representing standard filename of format "accident_<year>.csv.bz2"
}
\description{
This function generates a standardised filename string with csv.bz2 extension based on the year defined by the user
}
\note{
Function does not accept vectors containing multiple numerics
}
\examples{
make_filename(2014)

}
